+++ server/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Added:
+
+- recipe: support defining deprecated options
+- recipe: support `slapos.recipe.template` version 0.5
+
 ## [0.7.0] - 2022-01-06
 
 ### Added: